# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Paper Setting API",
    "description": "Local-first DOCX thesis formatting compiler",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "operationId": "create_job_api_v1_jobs_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_job_api_v1_jobs_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Delete Job",
        "operationId": "delete_job_api_v1_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/plan": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Plan",
        "operationId": "get_plan_api_v1_jobs__job_id__plan_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/template-plan": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Template Plan",
        "operationId": "get_template_plan_api_v1_jobs__job_id__template_plan_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TemplateCombinedPreview"
                    },
                    {
                      "$ref": "#/components/schemas/ManuscriptCompositionPreview"
                    }
                  ],
                  "title": "Response Get Template Plan Api V1 Jobs  Job Id  Template Plan Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/inspection": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Inspection",
        "operationId": "get_inspection_api_v1_jobs__job_id__inspection_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentInspection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/approve": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Approve Job",
        "operationId": "approve_job_api_v1_jobs__job_id__approve_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/report": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Report",
        "operationId": "get_report_api_v1_jobs__job_id__report_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/artifacts": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Artifacts",
        "operationId": "list_artifacts_api_v1_jobs__job_id__artifacts_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactView"
                  },
                  "title": "Response List Artifacts Api V1 Jobs  Job Id  Artifacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/artifacts/{artifact}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Download Artifact",
        "operationId": "download_artifact_api_v1_jobs__job_id__artifacts__artifact__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "artifact",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/events": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Stream Events",
        "operationId": "stream_events_api_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_create_job_api_v1_jobs_post": {
        "properties": {
          "document": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Document"
          },
          "rule_pack_id": {
            "type": "string",
            "title": "Rule Pack Id"
          },
          "render_preview": {
            "type": "boolean",
            "title": "Render Preview",
            "default": true
          },
          "mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "format",
                  "template"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "template_document": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Document"
          },
          "formatting_mode": {
            "type": "string",
            "enum": [
              "standardize",
              "preserve"
            ],
            "title": "Formatting Mode",
            "default": "standardize"
          }
        },
        "type": "object",
        "required": [
          "document",
          "rule_pack_id"
        ],
        "title": "Body_create_job_api_v1_jobs_post"
      },
...
      "DocumentInspection": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentItem"
            },
            "type": "array",
            "title": "Items"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionFormatSnapshot"
            },
            "type": "array",
            "title": "Sections"
          },
          "summary": {
            "$ref": "#/components/schemas/InspectionSummary"
          },
          "advanced": {
            "$ref": "#/components/schemas/AdvancedWordInspection"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "source_sha256",
          "source_filename",
          "items",
          "summary"
        ],
        "title": "DocumentInspection"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobView"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "JobPage"
      },
      "JobView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "current_stage": {
            "type": "string",
            "title": "Current Stage"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "template_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Filename"
          },
          "template_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Sha256"
          },
          "rule_pack_id": {
            "type": "string",
            "title": "Rule Pack Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "format",
              "template"
            ],
            "title": "Mode",
            "default": "format"
          },
          "formatting_mode": {
            "type": "string",
            "enum": [
              "standardize",
              "preserve"
            ],
            "title": "Formatting Mode",
            "default": "preserve"
          },
          "render_preview": {
            "type": "boolean",
            "title": "Render Preview"
          },
          "plan_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Version"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail"
          },
          "failure_count": {
            "type": "integer",
            "title": "Failure Count"
          },
          "max_attempts": {
            "type": "integer",
            "title": "Max Attempts"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "progress",
          "current_stage",
          "source_filename",
          "source_sha256",
          "rule_pack_id",
          "render_preview",
          "plan_version",
          "error_code",
          "error_detail",
          "failure_count",
          "max_attempts",
          "created_at",
          "updated_at"
        ],
        "title": "JobView"
      },
...
      "ManuscriptCompositionPreview": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "plan_kind": {
            "type": "string",
            "const": "manuscript_composition",
            "title": "Plan Kind",
            "default": "manuscript_composition"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "template_filename": {
            "type": "string",
            "title": "Template Filename"
          },
          "template_sha256": {
            "type": "string",
            "title": "Template Sha256"
          },
          "plan_version": {
            "type": "string",
            "title": "Plan Version"
          },
          "mappings": {
            "items": {
              "$ref": "#/components/schemas/ManuscriptFieldMapping"
            },
            "type": "array",
            "title": "Mappings"
          },
          "fill": {
            "$ref": "#/components/schemas/TemplateFillPreview"
          },
          "body_injection": {
            "$ref": "#/components/schemas/ManuscriptBodyInjectionPreview"
          },
          "structure": {
            "$ref": "#/components/schemas/TemplateStructurePreview"
          },
          "can_generate": {
            "type": "boolean",
            "title": "Can Generate"
          },
          "blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blockers"
          },
          "output_filename": {
            "type": "string",
            "title": "Output Filename"
          }
        },
        "type": "object",
        "required": [
          "source_filename",
          "source_sha256",
          "template_filename",
          "template_sha256",
          "plan_version",
          "mappings",
          "fill",
          "body_injection",
          "structure",
          "can_generate",
          "output_filename"
        ],
        "title": "ManuscriptCompositionPreview"
      },
...
      "PatchPlan": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "rule_pack_id": {
            "type": "string",
            "title": "Rule Pack Id"
          },
          "rule_pack_version": {
            "type": "string",
            "title": "Rule Pack Version"
          },
          "rule_pack_hash": {
            "type": "string",
            "title": "Rule Pack Hash"
          },
          "plan_version": {
            "type": "string",
            "title": "Plan Version"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/PatchOperation"
            },
            "type": "array",
            "title": "Operations"
          },
          "summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Summary"
          },
          "formatting_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FormattingPolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "notices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notices"
          },
          "uncovered_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uncovered Roles"
          }
        },
        "type": "object",
        "required": [
          "source_sha256",
          "rule_pack_id",
          "rule_pack_version",
          "rule_pack_hash",
          "plan_version",
          "operations",
          "summary"
        ],
        "title": "PatchPlan"
      },
...
      "TemplateCombinedPreview": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "plan_kind": {
            "type": "string",
            "const": "template_combined",
            "title": "Plan Kind",
            "default": "template_combined"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "plan_version": {
            "type": "string",
            "title": "Plan Version"
          },
          "fill": {
            "$ref": "#/components/schemas/TemplateFillPreview"
          },
          "structure": {
            "$ref": "#/components/schemas/TemplateStructurePreview"
          },
          "can_generate": {
            "type": "boolean",
            "title": "Can Generate"
          },
          "blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blockers"
          },
          "output_filename": {
            "type": "string",
            "title": "Output Filename"
          }
        },
        "type": "object",
        "required": [
          "source_filename",
          "source_sha256",
          "plan_version",
          "fill",
          "structure",
          "can_generate",
          "output_filename"
        ],
        "title": "TemplateCombinedPreview"
      },
...
    }
  }
}